requirements.txt
streamlit==1.40.2
anthropic==0.40.0
python-dotenv==1.0.1
chromadb==0.5.23
sentence-transformers==3.3.1
pypdf==5.1.0
python-docx==1.1.2